refactor: Simplify processing
This removes a bunch of unnecessary processing in special tags.
Also, it makes it "simple" to add components and default values within
them.

Still missing proper optimization of the `:c/for` tag, that was plain
wrong.

This should make it work, but in a bad way. The ideal implementation
would be one that avoids creating so many intermediary structures,
possibly not using the rucrsive function.